appディレクトに階層変更
