--- functions/src/data/accounts.ts
@@ -1,8 +1,0 @@
-/** 送信メール */
-export const sendMessage = {
-  form: 'さかのパーソナルサイト <support@saka-tai.com>',
-  subject: 'お問い合わせを受け付けました',
-  meSubject: 'お問い合わせがありました',
-  error: 'お問い合わせに失敗しました',
-  success: 'お問い合わせを受け付けました',
-} as const;
